requirements.txt
# Flask web app requirements for Azure App Service
Flask==3.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
pyodbc==5.2.0
langchain==0.3.26
langchain-community==0.3.26
google-generativeai==0.8.5
dateparser==1.2.1
gunicorn==21.2.0
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1
Werkzeug==3.1.3
MarkupSafe==3.0.2
blinker==1.9.0
email-validator==2.2.0
requests==2.32.4
flask-socketio==5.3.6
# For production WSGI server
# Gunicorn is recommended for running Flask in production
# Logging is handled by Python's logging module